test: refactor test list folder
