+++ django_forest/tests/resources/views/list/test_list_smart_fields.py
